Refactor: 접근제한자 설정 (#24)
